+++ src/main/java/com/example/examplemod/ExampleMod.java
@@ -6,6 +6,7 @@ import com.example.examplemod.blocks.ModBlocks;
 import com.example.examplemod.creativemodetab.ModCreativeModeTabs;
 import com.example.examplemod.data.ModDataComponents;
 import com.example.examplemod.item.ModItems;
+import com.example.examplemod.sound.ModSounds;
 import com.example.examplemod.stat.ModStats;
 import com.mojang.logging.LogUtils;
 
@@ -108,6 +109,8 @@ public class ExampleMod {
         ModDataComponents.register(modEventBus);
 
         ModStats.register(modEventBus);
+
+        ModSounds.register(modEventBus);
 
         // Register ourselves for server and other game events we are interested in.
         // Note that this is necessary if and only if we want *this* class (ExampleMod)