+++ src/main/java/com/example/examplemod/ExampleModDataGen.java
@@ -13,6 +13,7 @@ import com.example.examplemod.datagen.ModItemTagsProvider;
 import com.example.examplemod.datagen.ModModelProvider;
 import com.example.examplemod.datagen.ModPaintingTagsProvider;
 import com.example.examplemod.datagen.ModRecipeProvider;
+import com.example.examplemod.datagen.ModSoundsProvider;
 
 import net.minecraft.core.HolderLookup.Provider;
 import net.minecraft.world.level.storage.loot.parameters.LootContextParamSets;
@@ -44,5 +45,7 @@ public class ExampleModDataGen {
         generator.addProvider(true, new ModEquipmentAssetProvider(packOutput));
         generator.addProvider(true, new ModDataPackProvider(packOutput, lookupProvider));
         generator.addProvider(true, new ModPaintingTagsProvider(packOutput, lookupProvider));
+
+        generator.addProvider(true, new ModSoundsProvider(packOutput));
     }
 }